add StateWrapper
